AXIOM/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "AXIOM"
version = "1.0.0"
dependencies = [
    "typing_extensions>=4.0.0",
    "python-dotenv>=1.0.0",       # For env var loading
    "structlog>=24.0.0",          # Structured logging
    "psutil>=5.9.0",              # System monitoring (health checks)
]
requires-python = ">=3.11"
authors = [
  {name = "ZiaTechnica", website = "www.ziatechnica.org"},
]
maintainers = [
  {name = "Alejandro Rubio", email = "alejoserubio@gmail.com"}
]
description = "ARA's eXtented Intent & Execution Machine"
readme = "README.md"
keywords = ["Event", "Intent", "Bus", "Execution", "Fuck"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]
[project.optional-dependencies]
cli = ["rich", "click"]
dev = [
    "pytest>=8.0.0",
    "black",
    "isort",
    "pylint",
]